Add files via upload
This is the first installment, of a proposed expansion of music of DOC to cover musically underrepresented areas and to expand on an already amazing Soundtrack. This particular branch consists of 'New World' or Precolombian music. Music represents the indigenous cultures of North, Central and South America. The majority of the United States- Indian music comes from Civ IV Colonization and Carlos Nakai. However much of the Aztec music comes from a unique album from Ricardo Lozano that is the most accurate portrayal of Mesoamerican music to the best of my knowledge. To make this possible I have coded the creation of a New-World musical era so that the music should play for the Maya, Inca and Aztec between the classical and Rennaissance eras. Please give this a try and give feedback!