Reverted relation level base stabilities to the original 3/2/-2/-3 format. I was a bit overzealous when reducing them and ended up overcorrecting to the point that nations tended towards positive stabilities. From my testing, this makes stabilities tend around zero, with low negatives and positives common and high ones every now and then. Spain is easily the nation whos AI I most often see with high relations stability, though the Mongols and Arabia have had 6s on quite a few occasions in my testing.
In my opinion, I have sufficiently achieved my goal of creating an intuitive foreign relations stability system that is not as punishing as DOC's and does not give a great advantage to Catholic nations. Except Spain. AI Spain has a weirdly good ability to dominate the relations game.

While this means that the stability system is in a "gone gold" state, I will make adjustments in the future should they be needed or preferred.