Added a Turkic conquerors in Oman Added an independent Bahrein spawn to represent the Qarmatians under Abu Sa'id al-Jannabi Added barbarian spawns between 900 and 1000 AD around Bahrein to represent the Qarmatians under Abu Tahir al-Jannabi
So apparently Normal Maps are used to determine which cities flip when a Civ resurrects. I have yet to find the code for why this happens but I did some testing and changing Normal Maps does indeed change which cities are flipped in some instances of resurrection.

This should fix the issues with Nubia, Hungary and Ukraine, as well as similar but unreported issues with Burma, Chad, the Kazakhs, Oman, and Yemen