Add files via upload
This proposal meets between being authentic and appropriate for Dawn of Civilization. The majority of the movements are Guqin, as earlier requested, there are also flutes included in certain parts. This proposal is mostly Chinese but also includes a few Japanese, Korean and one Vietnamese track. All of the Korean tracks use the Gagageum instrument which has different notes, but sounds somewhat similar to the Guqin. Some of the tracks are historical pieces while others are covers of more recent songs. Songs from the newly invented Tank-Drum are included , and while technically not Ancient, have soothing sounds that should not distract the player and also have an Asian feel. 

Why this? 
In no way should this replace the existing Asian tracks, but this should give East Asia, a more filled out ambient experience for the player. The proposed music does not conflict with the existing music and indeed should provide enrichment and immersion. While not strictly historical the proposal as a whole looks to replicate the atmosphere of the Chinese Imperial era, and the courts of neighboring countries.  Dawn of Civilization would have nothing to lose by giving the region more aesthetic recognition.

This soundtrack is meant to run from the classical-renaissance eras as the existing tracks already do.

Thanks for considering!