African Music Module
I present here the African Music Module. Out of all of my four proposals this one was by far the most difficult to put together. Needless to say much of what could be considered traditional African music is singing and with noisy percussion effects may not be the best for DOC. I think with this selection I have come to some good compromises. Some of the music here is provided by the 'African Classical Music' ensemble which used traditional instruments from Ethiopia, others are authentic or (close to authentic' recreations of sounds from West/Central Africa. I am not sure about the suitable of these for DOC but considering the Civs that would be used for them (Mali and tropical Congo) I decided to show these anyway. In any case I think can provide ambience for the Sub Saharan Civilizations.

If nothing else, the tracks from the Classical African Music ensemble should be included! 

As the current DOC soundtrack does not have any representations of Africa to the best of my knowledge. Bringing some audio representation to the region is an important step. I realize it was almost two years when I discussed this, but I decided that now was no time was like the present to propose these. 

Thanks for considering these, and all of my other proposals!